# backend/openapi.yaml
openapi: 3.1.0
info:
  title: RAG Assistant Backend
  version: 0.2.1
paths:
...
  /api/upload:
    post:
      tags:
      - upload
      summary: Upload
      operationId: upload_api_upload_post
      parameters:
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Body_upload_api_upload_post'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/upload/op-status/{op_id}:
    get:
      tags:
      - upload
      summary: Op Status
      operationId: op_status_api_upload_op_status__op_id__get
      parameters:
      - name: op_id
        in: path
        required: true
        schema:
          type: string
          title: Op Id
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OpStatus'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    Body_upload_api_upload_post:
      properties:
        storeId:
          type: integer
          title: Storeid
          description: ID of the target store
        file:
          type: string
          format: binary
          title: File
          description: Document file to ingest
        displayName:
          anyOf:
          - type: string
            maxLength: 255
          - type: 'null'
          title: Displayname
          description: Optional display name override
      type: object
      required:
      - storeId
      - file
      title: Body_upload_api_upload_post
...
    DocumentStatus:
      type: string
      enum:
      - PENDING
      - RUNNING
      - DONE
      - ERROR
      title: DocumentStatus
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    OpStatus:
      properties:
        status:
          $ref: '#/components/schemas/DocumentStatus'
        error:
          anyOf:
          - type: string
          - type: 'null'
          title: Error
      type: object
      required:
      - status
      title: OpStatus
...
    UploadResponse:
      properties:
        op_id:
          type: string
          title: Op Id
        document_id:
          type: integer
          title: Document Id
        file_display_name:
          anyOf:
          - type: string
            maxLength: 255
          - type: 'null'
          title: File Display Name
        estimated_tokens:
          anyOf:
          - type: integer
          - type: 'null'
          title: Estimated Tokens
        estimated_cost_usd:
          anyOf:
          - type: number
          - type: 'null'
          title: Estimated Cost Usd
      type: object
      required:
      - op_id
      - document_id
      title: UploadResponse
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...